feat: changed loading strategy for existing games
feat: added support for future implementation of dynamic loading of localstorage data for GameBoard
